fix(sharing): consolidate share tab and ensure user keys are used
- consolidate files sharing sidebar tab components (no need for two
  sections we can re-use logic by just having one view).
- fix user shares: We reencrypt the root metadata with the new users
  key thus a new metadata key is generated and needs to be propagated
  to all subfolders.
