+++ src/services/filesSharingSection.ts
@@ -9,7 +9,7 @@ import { registerSidebarSection } from '@nextcloud/sharing/ui'
 import { defineAsyncComponent, defineCustomElement } from 'vue'
 
 /**
- * Register the filedrop section in the sharing sidebar for encrypted files.
+ * Register the user-, and filedrop section in the sharing sidebar for encrypted files.
  */
 export function registerSharingSidebarSection() {
 	const id = 'oca__end_to_end_encryption__sharing-sections'
@@ -26,21 +26,5 @@ export function registerSharingSidebarSection() {
 			return node.attributes['e2ee-is-encrypted'] === 1
 		},
 		element: id,
-	})
-
-	const idFiledrop = 'oca__end_to_end_encryption__sharing-sections-filedrop'
-	const FilesSharingSidebarSectionsFiledrop = defineCustomElement(
-		defineAsyncComponent(() => import('../views/FilesSharingSidebarSectionFiledrop.vue')),
-		{ shadowRoot: false },
-	)
-	window.customElements.define(idFiledrop, FilesSharingSidebarSectionsFiledrop)
-
-	registerSidebarSection({
-		id: 'end_to_end_encryption:filedrop',
-		order: 55,
-		enabled(node: INode) {
-			return node.attributes['e2ee-is-encrypted'] === 1
-		},
-		element: idFiledrop,
 	})
 }
